fix(models): keep capability resolver within file budget
