+++ src/lib/modelCapabilities.ts
@@ -12,12 +12,7 @@ import {
 } from "@/shared/constants/modelSpecs";
 import { getSyncedCapability } from "@/lib/modelsDevSync";
 import { MODELS_DEV_PROVIDER_MAP } from "@/lib/modelsDevSync/transform";
-import {
-  getModelContextOverride,
-  getModelContextOverrideRecord,
-  type ModelContextOverride,
-} from "@/lib/db/modelContextOverrides";
-import { getModelCapabilityOverride } from "@/lib/db/modelCapabilityOverrides";
+import { getModelContextOverride } from "@/lib/db/modelContextOverrides";
 import { getCustomModelVisionOverride } from "@/lib/db/models";
 import type { ModelCapabilityResolutionSnapshot } from "@/lib/modelCapabilityResolutionSnapshot";
 import {
@@ -25,16 +20,22 @@ import {
   resolveClampedMaxInputLimit,
   type ResolvedLimitSource,
 } from "@/lib/modelCapabilityLimits";
-import { resolveAudioCapability, resolveVideoCapability } from "@/lib/modelCapabilityModalities";
 import {
-  isResolutionSnapshot,
-  resolveClampedMaxInputLimit,
-  type ResolvedLimitSource,
-} from "@/lib/modelCapabilityLimits";
+  getContextOverride,
+  getContextOverrideRecord,
+  getInputTokenCapabilityOverride,
+  getOutputTokenCapabilityOverride,
+} from "@/lib/modelCapabilityOverrideResolution";
+import { resolveAudioCapability, resolveVideoCapability } from "@/lib/modelCapabilityModalities";
 export type { ModelCapabilityResolutionSnapshot } from "@/lib/modelCapabilityResolutionSnapshot";
 export type { ResolvedLimitSource } from "@/lib/modelCapabilityLimits";
 export { isPersistedResolvedLimitSource } from "@/lib/modelCapabilityLimits";
 export { createModelCapabilityResolutionSnapshot } from "@/lib/modelCapabilityResolutionSnapshot";
+
+/** Resolve a persisted context override by canonical id, then exact raw alias. */
+export function getResolvedModelContextOverride(input: CapabilityInput): number | null {
+  return getContextOverride(resolveCapabilityInput(input));
+}
 export { resolveAudioCapability } from "@/lib/modelCapabilityModalities";
 import { isVisionModelId } from "@/shared/constants/visionModels";
 import { getUnsupportedParams } from "@omniroute/open-sse/config/providerRegistry.ts";
@@ -536,141 +537,6 @@ function resolveVisionCapability(
   if (modelIdLikelyVision(modelId)) return true;
 
   return null;
-}
-
-/**
- * Issue #6524: an operator-set `max_output_tokens` capability override (see
- * `src/lib/db/modelCapabilityOverrides.ts`) is the manual escape hatch for a
- * wrong/stale synced `limit_output` value (e.g. a provider's models.dev catalog
- * row reporting `limit_output` equal to `limit_context`). It already won over the
- * synced value in `getResolvedModelCapabilities().maxOutputTokens` — this helper
- * makes `getExplicitModelOutputCap()` (used by the reasoning-token-buffer clamp)
- * consult the same override so both read paths agree.
- */
-/**
- * Exact-match capability override lookup with intentional raw-alias fallback.
- *
- * An override may be stored under either the canonical model id or the exact
- * provider-scoped raw alias the operator used (e.g. `github/claude-opus-4.5`
- * resolving to canonical `claude-opus-4-5-20251101`). We consult the canonical
- * id first, then the raw alias — both are exact provider/model matches. There is
- * deliberately NO suffix/effort/family inheritance: an override for
- * `codex/gpt-5.6` never applies to `codex/gpt-5.6-high`.
- */
-function getCapabilityOverride(
-  resolved: { provider: string | null; model: string | null; rawModel: string | null },
-  key: "max_input_tokens" | "max_output_tokens",
-  bulkOverrides?: ReadonlyMap<string, ReadonlyMap<string, number>> | null
-): number | null {
-  const canonical = getModelCapabilityOverride(
-    resolved.provider,
-    resolved.model,
-    key,
-    bulkOverrides
-  );
-  if (canonical !== null) return canonical;
-  return resolved.rawModel && resolved.rawModel !== resolved.model
-    ? getModelCapabilityOverride(resolved.provider, resolved.rawModel, key, bulkOverrides)
-    : null;
-}
-
-function getContextOverrideRecord(
-  resolved: {
-    provider: string | null;
-    model: string | null;
-    rawModel: string | null;
-  },
-  snapshot?: ModelCapabilityResolutionSnapshot | null
-): ModelContextOverride | null {
-  const lookup = (model: string | null) => {
-    if (!resolved.provider || !model) return null;
-    if (!snapshot) return getModelContextOverrideRecord(resolved.provider, model);
-
-    const record = snapshot.contextOverrideRecords?.get(resolved.provider)?.get(model);
-    if (record) return record;
-
-    const legacyContext = snapshot.contextOverrides.get(resolved.provider)?.get(model);
-    return legacyContext === undefined
-      ? null
-      : {
-          provider: resolved.provider,
-          modelId: model,
-          realContext: legacyContext,
-          source: "manual" as const,
-          refreshedAt: "",
-        };
-  };
-  const canonical = lookup(resolved.model);
-  if (canonical) return canonical;
-  return resolved.rawModel && resolved.rawModel !== resolved.model
-    ? lookup(resolved.rawModel)
-    : null;
-}
-
-function getContextOverride(
-  resolved: {
-    provider: string | null;
-    model: string | null;
-    rawModel: string | null;
-  },
-  snapshot?: ModelCapabilityResolutionSnapshot | null
-): number | null {
-  return getContextOverrideRecord(resolved, snapshot)?.realContext ?? null;
-}
-
-/**
- * Resolve a persisted context override by canonical id, then by the exact raw
- * alias supplied by the caller. Neither lookup inherits to related models.
- *
- * `snapshot` is the #9147 build-local bulk load; when supplied the on-demand
- * SQLite read is skipped and the preloaded nested map is used instead.
- */
-export function getResolvedModelContextOverride(
-  input: CapabilityInput,
-  snapshot?: ModelCapabilityResolutionSnapshot | null
-): number | null {
-  return getContextOverride(resolveCapabilityInput(input), snapshot);
-}
-
-function getInputTokenCapabilityOverride(
-  resolved: { provider: string | null; model: string | null; rawModel: string | null },
-  snapshot?: ModelCapabilityResolutionSnapshot | null
-): number | null {
-  return getCapabilityOverride(
-    resolved,
-    "max_input_tokens",
-    snapshot ? (snapshot.maxInputTokenOverrides ?? new Map()) : undefined
-  );
-}
-
-function getOutputTokenCapabilityOverride(
-  resolved: { provider: string | null; model: string | null; rawModel: string | null },
-  snapshot?: ModelCapabilityResolutionSnapshot | null
-): number | null {
-  if (!snapshot) return getCapabilityOverride(resolved, "max_output_tokens");
-
-  const current = snapshot.maxTokenOverrides ?? new Map();
-  const historical = snapshot.maxOutputTokenOverrides ?? new Map();
-  const canonical =
-    getModelCapabilityOverride(resolved.provider, resolved.model, "max_output_tokens", current) ??
-    getModelCapabilityOverride(resolved.provider, resolved.model, "max_output_tokens", historical);
-  if (canonical !== null) return canonical;
-
-  if (!resolved.rawModel || resolved.rawModel === resolved.model) return null;
-  return (
-    getModelCapabilityOverride(
-      resolved.provider,
-      resolved.rawModel,
-      "max_output_tokens",
-      current
-    ) ??
-    getModelCapabilityOverride(
-      resolved.provider,
-      resolved.rawModel,
-      "max_output_tokens",
-      historical
-    )
-  );
 }
 
 export function getExplicitModelOutputCap(
